move renderer into it's own file
